requirement.txt
mysql-connector-python
psycopg2-binary
Faker